# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple API for managing player profiles and achievements",
        "title": "Player Profile API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/achievements": {
            "get": {
                "description": "Get all achievements with pagination, default page is 1 and default pageSize is 10",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Get all achievements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/response.AchievementResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new achievement with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Create a new achievement",
                "parameters": [
                    {
                        "description": "Create Achievement Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/achievements/{achievementID}": {
            "get": {
                "description": "Get an achievement by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Get an achievement by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Achievement ID",
                        "name": "achievementID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an achievement with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Update an achievement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Achievement ID",
                        "name": "achievementID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Achievement Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateAchievementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an achievement by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Delete an achievement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Achievement ID",
                        "name": "achievementID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/achievements/{achievementID}/players": {
            "get": {
                "description": "Get an achievement with players by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Achievement"
                ],
                "summary": "Get an achievement with players",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Achievement ID",
                        "name": "achievementID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.CreateAchievementRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Achievement description",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "name": {
                    "description": "Achievement name",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                }
            }
        },
...
        "request.UpdateAchievementRequest": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Achievement description",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "name": {
                    "description": "Achievement name",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                }
            }
        },
...
        "response.AchievementResponse": {
            "type": "object",
            "required": [
                "description",
                "id",
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Achievement description",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "id": {
                    "description": "Achievement ID (primary key) in the database",
                    "type": "integer"
                },
                "name": {
                    "description": "Achievement name",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
        "response.BaseResponse": {
            "description": "Base response structure",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP status code of the response",
                    "type": "integer",
                    "x-order": "0",
                    "example": 200
                },
                "status": {
                    "description": "Status of the response",
                    "type": "string",
                    "x-order": "1",
                    "example": "success"
                },
                "message": {
                    "description": "Message of the response",
                    "type": "string",
                    "x-order": "2",
                    "example": "Operation completed successfully"
                },
                "data": {
                    "description": "Data payload of the response",
                    "x-order": "3"
                }
            }
        }
    }
}